+++ database/migrations/2025_04_10_072345_update_settings_table_add_taxpercent_column.php
@@ -25,7 +25,13 @@ return new class extends Migration
     {
         Schema::table('settings', function (Blueprint $table) {
             //drop the column tax_percentage
-            $table->dropColumn('tax_percentage');
+            if (Schema::hasColumn('settings', 'tax_percentage')) {
+                $table->dropColumn('tax_percentage');
+            }
+            //drop the column tax_threshold
+            if (Schema::hasColumn('settings', 'tax_threshold')) {
+                $table->dropColumn('tax_threshold');
+            }
         });
     }
 };
